requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
anthropic==0.37.0
supabase==2.9.0
python-dotenv==1.0.1
httpx==0.27.2
aiogram==3.13.0
pydantic==2.9.2
